feat: renaming, change structure and add categories to Book model
