+++ src/books/books.service.ts
@@ -1,7 +1,7 @@
 import { InjectModel } from '@nestjs/mongoose';
 import { Book } from '../schemas/Book/book.schema';
 import { Model } from 'mongoose';
-import { CreateBookDTO } from './dto/createBook.dto';
+import { CreateBookDTO } from './dto/create-book.dto';
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
@@ -13,11 +13,11 @@ export class BooksService {
     return createdBook.save();
   }
 
-  async findAll(): Promise<Book[]> {
+  async findAllBooks(): Promise<Book[]> {
     return this.bookModel.find().exec();
   }
 
-  async findOne(bookId: string): Promise<Book | null> {
+  async findOneBook(bookId: string): Promise<Book | null> {
     return this.bookModel.findOne({ _id: bookId });
   }
 }
